+++ src/helper/DetermineTime.cpp
@@ -1,14 +1,12 @@
 #include "DetermineTime.h"
 #include <chrono>
 
-// Obtiene el tiempo actual en milisegundos
-long long DetermineTime::getCurrentMillisecondsTime() {
+long DetermineTime::getCurrentMillisecondsTime() {
     auto now = std::chrono::system_clock::now();
     auto duration = std::chrono::duration_cast<std::chrono::milliseconds>(now.time_since_epoch());
     return duration.count();
 }
 
-// Calcula el tiempo transcurrido en milisegundos desde initTime
-long long DetermineTime::getMillisecondsPassed(long long initTime) {
+long DetermineTime::getMillisecondsPassed(long initTime) {
     return getCurrentMillisecondsTime() - initTime;
 }
